Refactor plotting functionality into separate submodule and tweak error handling
